refactor(web): extract <ItemDetail> component with embedded prop (TASK-2106)
Move the entire item-content <script>+template+<style> out of the
[collection]/[slug] route into a shared, embeddable
web/src/lib/components/items/ItemDetail.svelte. The full-page route
becomes a thin wrapper that renders <ItemDetail> with embedded=false,
preserving byte-identical full-page behavior.

Foundation task of PLAN-2105 (Asana-style split-pane navigation). No
user-visible change: nothing mounts embedded yet (TASK-2112). The
embedded=true paths are implemented so they compile and are correct,
gated behind the `embedded` prop.

- Props {username, wsSlug, collSlug, ref, embedded=false, onClose,
  onGone, onNavigateAway, onReady} replace page.params reads. itemSlug
  derives from `ref`, re-driving loadData + the collabKey chain on ref
  change without a remount.
- Route-away goto's parameterized as onGone/onNavigateAway with
  embedded=false defaults that reproduce today's exact goto's (full
  pane callback wiring is TASK-2107).
- embedded behaviors (gated, not yet mounted): page-title deferral,
  ?graph drawer driven by local $state instead of the URL param, ?new
  never read, breadcrumb hidden.
- The whole collab machinery (provider $effect mint+destroy incl.
  doc.destroy(), flusher, Editor {#key}+{#if ydoc}, editorInstance-null
  $effect, SSE/sync onMount/onDestroy subs, beforeunload flush) stays
  in one component boundary inside ItemDetail.
- `export const snapshot` stays on the route wrapper (a SvelteKit
  route-module concern); ItemDetail threads scroll-restoration
  readiness back via onReady so the wrapper's snapshot parks until the
  child's item matches the URL.

schemaVersion.ts NOT bumped (no ProseMirror node-spec change).
